\docType{data}
\name{mRNAExpression}
\alias{mRNAExpression}
\title{mRNA expression data from TCGA - KIRC}
\format{A data frame
}

\usage{
mRNAExpression
}
\description{
mRNA expression of 1009 genes expressed under 534 samples
}
\examples{
data("mRNAExpression")
}
\references{
Thierry Chekouo and Himadri Mukherjee (2023), \emph{A Bayesian Hierarchical Hidden Markov Model for clustering and gene selection: Application to Kidney cancer gene expression data}, \emph{submitted}.
}
\keyword{RealData}
