\docType{data}
\name{SimilarityBP}
\alias{SimilarityBP}
\title{Similarities between genes}
\format{A data frame}

\usage{
SimilarityBP
}
\description{
Similarities between genes computed using the Biological Process in GO
}
\examples{
data("SimilarityBP")
}
\references{
Thierry Chekouo and Himadri Mukherjee (2023), \emph{A Bayesian Hierarchical Hidden Markov Model for clustering and gene selection: Application to Kidney cancer gene expression data}, \emph{submitted}.
}
\keyword{RealData}
